refactor: isolated tests
