+++ src/__test__/presets/routes/admin.ts
@@ -1,24 +1,24 @@
-import { createAdminRoleIfNotExist } from "../../createAdminRoleIfNotExists";
-import { createClientRoleIfNotExist } from "../../createClientRoleIfNotExists";
-import { createMemberRoleIfNotExist } from "../../createMemberRoleIfNotExists";
 import { createAdmin } from "@repositories/user/admin";
 import { createClient } from "@repositories/user/client";
 import { createMember } from "@repositories/user/member";
 import { prismaClient } from "@libs/prisma";
 import { Admin, Client, Member, Role, User } from "@prisma/client";
+import { createAdminRoleIfNotExist } from "../createAdminRoleIfNotExists";
+import { createClientRoleIfNotExist } from "../createClientRoleIfNotExists";
+import { createMemberRoleIfNotExist } from "../createMemberRoleIfNotExists";
 
 const createAdminToAuthenticate = {
-  name: "Test Admin Authenticate to member crud test",
+  name: "Test Admin Authenticate to admin crud test",
   password: "123",
 };
 
 const createClientToAuthenticate = {
-  name: "Test Client Authenticate to member crud test",
+  name: "Test Client Authenticate to admin crud test",
   password: "123",
 };
 
 const createMemberToAuthenticate = {
-  name: "Test Member Authenticate to member crud test",
+  name: "Test Member Authenticate to admin crud test",
   password: "123",
 };
 
@@ -26,7 +26,7 @@ let userAdmin: User & { role: Role; admin: Admin };
 let userClient: User & { role: Role; client: Client };
 let userMember: User & { role: Role; member: Member };
 
-export const presetToMemberTests = async () => {
+export const presetToAdminTests = async () => {
   const RoleAdminId = await createAdminRoleIfNotExist();
   const RoleClientId = await createClientRoleIfNotExist();
   const RoleMemberId = await createMemberRoleIfNotExist();
@@ -38,7 +38,7 @@ export const presetToMemberTests = async () => {
 
   userClient = await createClient({
     ...createClientToAuthenticate,
-    address: { house: "3", square: "3" },
+    address: { house: "0", square: "0" },
     roleId: RoleClientId,
   });
 
@@ -51,7 +51,7 @@ export const presetToMemberTests = async () => {
   return { userAdmin, userClient, userMember };
 };
 
-export const cleanMemberTestDatabase = async () => {
+export const cleanAdminTestDatabase = async () => {
   await prismaClient.user.deleteMany({
     where: {
       id: {
