--- src/__test__/utils/users.ts
@@ -1,14 +1,0 @@
-export const userAsAdmin = {
-  name: "Test Admin",
-  password: "123",
-};
-
-export const userAsClient = {
-  name: "Test Client",
-  password: "123",
-};
-
-export const userAsMember = {
-  name: "Test Member",
-  password: "123",
-};
